Do a large refactor to put most of the code into modules
The modules directory contains configuration that is the same across
different environments. When we need to update something, we make
changes to the module and then populate the changes to the environments.

Variables in terraform.tfvars are extra-thin, meaning that they contain
only the things that will be different accross environments.